requirements.txt
# Azure Functions Core
azure-functions>=1.18.0

# Azure Services
azure-storage-blob>=12.19.0
azure-identity>=1.15.0
azure-core>=1.29.0
azure-search-documents>=11.4.0

# Database
pyodbc>=5.0.1
aiosqlite>=0.19.0

# AI/OpenAI
openai>=1.12.0

# HTTP and File Handling
python-multipart>=0.0.6
requests>=2.31.0

# Data Models and Validation
pydantic>=2.5.0

# Configuration and Environment
python-dotenv>=1.0.0

# Document Processing
python-docx>=1.1.0
PyPDF2>=3.0.1

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0